Update tests for mutlisigFreezeGuard
